modified multiple files

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 	long long int ll;
 	float f;
 
-	printf("size of a Char: %lu byte(s)\n", (unsigned long)sizeof(c));
-	printf("size of an int: %lu byte(s)\n", (unsigned long)sizeof(i));
-	printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(l));
-	printf("size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(ll));
-	printf("size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
+	printf("size of a Char: %lu byte(s)\n", sizeof(c));
+	printf("size of an int: %lu byte(s)\n", sizeof(i));
+	printf("size of a long int: %lu byte(s)\n", sizeof(l));
+	printf("size of a long long int: %lu byte(s)\n", sizeof(ll));
+	printf("size of a float: %lu byte(s)\n", sizeof(f));
 	return (0);
 }
